refactor: programTypes and scopeTypes
