+++ src/core_modules/capture-core/hooks/useProgramInfo/programTypes.const.js
@@ -1,5 +1,2 @@
 // @flow
-export const programTypes = {
-    EVENT_PROGRAM: 'eventProgram',
-    TRACKER_PROGRAM: 'trackerProgram',
-};
+export { programTypes } from '../../metaData';
