+++ src/core_modules/capture-core/metaData/helpers/constants/index.js
@@ -1,0 +1,3 @@
+// @flow
+export { programTypes } from './programTypes.const';
+export { scopeTypes } from './scopeTypes.const';
