+++ src/core_modules/capture-core/metaData/helpers/constants/programTypes.const.js
@@ -1,0 +1,5 @@
+// @flow
+export const programTypes = {
+    EVENT_PROGRAM: 'EVENT_PROGRAM',
+    TRACKER_PROGRAM: 'TRACKER_PROGRAM',
+};
