+++ src/core_modules/capture-core/metaData/helpers/constants/scopeTypes.const.js
@@ -1,0 +1,7 @@
+// @flow
+import { programTypes } from './programTypes.const';
+
+export const scopeTypes = {
+    ...programTypes,
+    TRACKED_ENTITY_TYPE: 'TRACKED_ENTITY_TYPE',
+};
