+++ src/core_modules/capture-core/metaData/helpers/getScopeFromScopeId.js
@@ -3,12 +3,6 @@ import { EventProgram, TrackerProgram } from '../Program';
 import { TrackedEntityType } from '../TrackedEntityType';
 import { programCollection, trackedEntityTypesCollection } from '../../metaDataMemoryStores';
 
-export const scopeTypes = {
-    TRACKER_PROGRAM: 'TRACKER_PROGRAM',
-    EVENT_PROGRAM: 'EVENT_PROGRAM',
-    TRACKED_ENTITY_TYPE: 'TRACKED_ENTITY_TYPE',
-};
-
 type Scope = (EventProgram | TrackerProgram | TrackedEntityType)
 
 export function getScopeFromScopeId(scopeId: ?string): ?Scope {
