+++ src/core_modules/capture-core/metaData/helpers/index.js
@@ -16,3 +16,4 @@ export {
 } from './trackedEntityType/getTrackedEntityTypeThrowIfNotFound';
 export { convertValues as convertDataElementsValues } from './DataElements';
 export { getScopeFromScopeId } from './getScopeFromScopeId';
+export { programTypes, scopeTypes } from './constants';
